# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Crypto Trading Journal API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: trades
    description: Trade journal operations
  - name: watchlist
    description: Watchlist operations
  - name: analytics
    description: Analytics and reporting
  - name: settings
    description: User settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List all trades
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: setupType
          in: query
          schema:
            type: string
        - name: coin
          in: query
          schema:
            type: string
        - name: outcome
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Create and score a new trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Created trade with score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"

  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a single trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found
    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update a trade (outcome, notes, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeUpdate"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Delete a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /trades/score:
    post:
      operationId: scoreTrade
      tags: [trades]
      summary: Score trade inputs without saving (live preview)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "200":
          description: Calculated score output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeScore"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [analytics]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /analytics/equity-curve:
    get:
      operationId: getEquityCurve
      tags: [analytics]
      summary: Get equity curve data points
      responses:
        "200":
          description: Equity curve
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquityPoint"

  /analytics/by-setup:
    get:
      operationId: getAnalyticsBySetup
      tags: [analytics]
      summary: Win rate and PnL grouped by setup type
      responses:
        "200":
          description: Analytics by setup
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SetupAnalytic"

  /analytics/by-coin:
    get:
      operationId: getAnalyticsByCoin
      tags: [analytics]
      summary: PnL and win rate grouped by coin
      responses:
        "200":
          description: Analytics by coin
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoinAnalytic"

  /analytics/mistakes:
    get:
      operationId: getMistakeFrequency
      tags: [analytics]
      summary: Mistake frequency tracking
      responses:
        "200":
          description: Mistake data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MistakeEntry"

  /watchlist:
    get:
      operationId: listWatchlist
      tags: [watchlist]
      summary: List watchlist items
      responses:
        "200":
          description: Watchlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
    post:
      operationId: createWatchlistItem
      tags: [watchlist]
      summary: Add item to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"

  /watchlist/{id}:
    patch:
      operationId: updateWatchlistItem
      tags: [watchlist]
      summary: Update watchlist item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
    delete:
      operationId: deleteWatchlistItem
      tags: [watchlist]
      summary: Delete watchlist item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get user settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    TradeInput:
      type: object
      required:
        - coin
        - setupType
        - timeframe
        - btcCondition
        - altCondition
        - narrativeStrength
        - levelClarity
        - timeframeAlignment
        - retestQuality
        - volumeStrength
        - candleImpulse
        - followThrough
        - stopLossPct
        - tp1Pct
        - tp2Pct
        - entryDistance
        - spaceToResistance
        - rrQuality
        - overextension
        - eventRisk
        - liquidityRisk
      properties:
        coin:
          type: string
        setupType:
          type: string
        timeframe:
          type: string
        btcCondition:
          type: string
        altCondition:
          type: string
        narrativeStrength:
          type: string
        levelClarity:
          type: string
        timeframeAlignment:
          type: string
        retestQuality:
          type: string
        volumeStrength:
          type: string
        candleImpulse:
          type: string
        followThrough:
          type: string
        stopLossPct:
          type: number
        tp1Pct:
          type: number
        tp2Pct:
          type: number
        entryDistance:
          type: string
        spaceToResistance:
          type: string
        rrQuality:
          type: string
        overextension:
          type: string
        eventRisk:
          type: string
        liquidityRisk:
          type: string
        notes:
          type: ["string", "null"]
        mode:
          type: string
          enum: [trade, watchlist]

    TradeUpdate:
      type: object
      properties:
        outcome:
          type: ["string", "null"]
        actualPnlPct:
          type: ["number", "null"]
        exitPrice:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        mistakeTags:
          type: ["string", "null"]
        status:
          type: ["string", "null"]

    Trade:
      type: object
      required:
        - id
        - coin
        - setupType
        - timeframe
        - btcCondition
        - altCondition
        - narrativeStrength
        - levelClarity
        - timeframeAlignment
        - retestQuality
        - volumeStrength
        - candleImpulse
        - followThrough
        - stopLossPct
        - tp1Pct
        - tp2Pct
        - entryDistance
        - spaceToResistance
        - rrQuality
        - overextension
        - eventRisk
        - liquidityRisk
        - finalScore
        - tradeStatus
        - suggestedAllocationPct
        - suggestedSlPct
        - suggestedTpStructure
        - suggestedRr
        - tradeWarnings
        - calculatedRisk
        - expectedProfitPct
        - expectedLossPct
        - finalDecision
        - createdAt
      properties:
        id:
          type: integer
        coin:
          type: string
        setupType:
          type: string
        timeframe:
          type: string
        btcCondition:
          type: string
        altCondition:
          type: string
        narrativeStrength:
          type: string
        levelClarity:
          type: string
        timeframeAlignment:
          type: string
        retestQuality:
          type: string
        volumeStrength:
          type: string
        candleImpulse:
          type: string
        followThrough:
          type: string
        stopLossPct:
          type: number
        tp1Pct:
          type: number
        tp2Pct:
          type: number
        entryDistance:
          type: string
        spaceToResistance:
          type: string
        rrQuality:
          type: string
        overextension:
          type: string
        eventRisk:
          type: string
        liquidityRisk:
          type: string
        notes:
          type: ["string", "null"]
        mode:
          type: string
        outcome:
          type: ["string", "null"]
        actualPnlPct:
          type: ["number", "null"]
        exitPrice:
          type: ["number", "null"]
        mistakeTags:
          type: ["string", "null"]
        status:
          type: string
        finalScore:
          type: number
        tradeStatus:
          type: string
        suggestedAllocationPct:
          type: number
        suggestedSlPct:
          type: number
        suggestedTpStructure:
          type: string
        suggestedRr:
          type: number
        tradeWarnings:
          type: string
        calculatedRisk:
          type: number
        expectedProfitPct:
          type: number
        expectedLossPct:
          type: number
        finalDecision:
          type: string
        createdAt:
          type: string

    TradeScore:
      type: object
      required:
        - finalScore
        - tradeStatus
        - suggestedAllocationPct
        - suggestedSlPct
        - suggestedTpStructure
        - suggestedRr
        - tradeWarnings
        - calculatedRisk
        - expectedProfitPct
        - expectedLossPct
        - finalDecision
      properties:
        finalScore:
          type: number
        tradeStatus:
          type: string
        suggestedAllocationPct:
          type: number
        suggestedSlPct:
          type: number
        suggestedTpStructure:
          type: string
        suggestedRr:
          type: number
        tradeWarnings:
          type: string
        calculatedRisk:
          type: number
        expectedProfitPct:
          type: number
        expectedLossPct:
          type: number
        finalDecision:
          type: string

    DashboardStats:
      type: object
      required:
        - totalCapital
        - totalPnlPct
        - winRate
        - avgRr
        - tradeCount
        - openTradesCount
        - winCount
        - lossCount
        - recentTrades
      properties:
        totalCapital:
          type: number
        totalPnlPct:
          type: number
        winRate:
          type: number
        avgRr:
          type: number
        tradeCount:
          type: integer
        openTradesCount:
          type: integer
        winCount:
          type: integer
        lossCount:
          type: integer
        recentTrades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"

    EquityPoint:
      type: object
      required:
        - date
        - equity
        - tradeId
      properties:
        date:
          type: string
        equity:
          type: number
        tradeId:
          type: integer

    SetupAnalytic:
      type: object
      required:
        - setupType
        - tradeCount
        - winRate
        - totalPnlPct
        - avgScore
      properties:
        setupType:
          type: string
        tradeCount:
          type: integer
        winRate:
          type: number
        totalPnlPct:
          type: number
        avgScore:
          type: number

    CoinAnalytic:
      type: object
      required:
        - coin
        - tradeCount
        - winRate
        - totalPnlPct
      properties:
        coin:
          type: string
        tradeCount:
          type: integer
        winRate:
          type: number
        totalPnlPct:
          type: number

    MistakeEntry:
      type: object
      required:
        - tag
        - count
      properties:
        tag:
          type: string
        count:
          type: integer

    WatchlistItem:
      type: object
      required:
        - id
        - coin
        - setupType
        - timeframe
        - notes
        - createdAt
      properties:
        id:
          type: integer
        coin:
          type: string
        setupType:
          type: string
        timeframe:
          type: string
        notes:
          type: string
        outcome:
          type: ["string", "null"]
        createdAt:
          type: string

    WatchlistItemInput:
      type: object
      required:
        - coin
        - setupType
        - timeframe
        - notes
      properties:
        coin:
          type: string
        setupType:
          type: string
        timeframe:
          type: string
        notes:
          type: string

    WatchlistItemUpdate:
      type: object
      properties:
        notes:
          type: string
        outcome:
          type: string

    Settings:
      type: object
      required:
        - id
        - totalCapital
        - riskProfilePct
        - defaultRiskPct
        - maxAllocationPct
      properties:
        id:
          type: integer
        totalCapital:
          type: number
        riskProfilePct:
          type: number
        defaultRiskPct:
          type: number
        maxAllocationPct:
          type: number

    SettingsUpdate:
      type: object
      properties:
        totalCapital:
          type: number
        riskProfilePct:
          type: number
        defaultRiskPct:
          type: number
        maxAllocationPct:
          type: number