# lib/api-zod/src/generated/types/coinAnalytic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface CoinAnalytic {
  coin: string;
  tradeCount: number;
  winRate: number;
  totalPnlPct: number;
}