# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Journal API
 * OpenAPI spec version: 0.1.0
 */
import type { Trade } from "./trade";

export interface DashboardStats {
  totalCapital: number;
  totalPnlPct: number;
  winRate: number;
  avgRr: number;
  tradeCount: number;
  openTradesCount: number;
  winCount: number;
  lossCount: number;
  recentTrades: Trade[];
}