# lib/api-zod/src/generated/types/equityPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface EquityPoint {
  date: string;
  equity: number;
  tradeId: number;
}